Use an initial version of 0.1.0 for plugins/plugin api
Instead of starting with a bold 1.0.0 for the initial plugins api
version, we're going to start with a more reasonable 0.1.0. While
we're at it, let's use an initial version of 0.1.0 for all plugin
versions.

After this, the plugin versions will progress independently from each
other and from the api version.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>