Be more strict about missing fields
The biggest driver for this is to ensure that a user can write a
clause "and not ct.error exists" and have the clause evaluate to true
if there is no error. Previously, ct.error would return an empty
string, which still counted as "existing".

Instead of converting missing fields in the json object to "", be more
strict about missing fields:

- In Next(), if the eventTime or eventType are missing, skip the event
  by returning timeout.

- In getUser(), change the return value to (present, user) and set
  present to false if userIdentity is missing or if the code can't
  identify a user. In get_fields() for ct.user, set field_present to
  false if getUser()'s present is false. In event_to_string, use a
  blank string for the user if not present.

- In event_to_string, if eventSource, awsRegion, or eventName are
  missing, return an error string.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>