refactor(dummy_c): avoid reallocating event mem
This commit is intened to work with the scap patch attached to https://github.com/falcosecurity/plugin-sdk-go/pull/10
Since with that patch, scap will not free
event memory anymore, we have to manage it inside the plugin.
Moreover, we can reuse the same struct and data buffer, instead of allocating new memory on each next() call.

Co-authored-by: Jason Dellaluce <jasondellaluce@gmail.com>
Signed-off-by: Leonardo Grasso <me@leonardograsso.com>