C++ SDK for Source Plugins
To make it easier for potential plugin authors to write plugins in
C++, this is a SDK that provides base C++ objects that manage many
details of the plugins interface.

Plugin authors can create classes that derive from
falcosecurity::source_plugin and falcosecurity::plugin_instance
and implement the methods in falcosecurity::source_plugin_iface
and falcosecurity::plugin_instance_iface to implement a source
plugin.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>