Fix off-by-one error in NextBatch
NextBatch is presented a long array of sdk.EventWriters and associates
events with them until the first error (which could include
eof/timeout).

If the last event results in an error, the loop would increment n,
then check the error, then return. n would be one bigger than
expected.

Fix this by checking for error + breaking inside the loop instead of
checking for error in the loop condition.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>