update(plugins): bump SDK Go version
This points to a specific commit hash to include recent changes not-yet released in the Go SDK. For the same reason,
the version of each plugin is not bumped, thus not meant to be tagged and released after this. The goal here
is to create dev builds that are able to comply with the new 1.0.0 Plugin API version.

Signed-off-by: Jason Dellaluce <jasondellaluce@gmail.com>