refactor(plugins/k8saudit): do not return <NA> on missing field values
In the legacy implementation, list values used to include the value <NA> whenever a given field could
not be extracted (see: https://github.com/falcosecurity/falco/pull/1601). However, this makes the semantics
of missing values ambigue and not coherent. Instead, here we just skip missing values. If a value is not found
when extracted, the plugin framework is able to signal the "field is not present" info. For list values, missing
values are simply skipped now. This works as expected with operators such as `in` and `intersects`, for which
the <NA> string was ambiguous anyways. If a list field is accessed with an index argument, then the extracted
value can either be a list with a single value (the one actually extracted at the positional index of the arg),
or an empty string (which is interpreted by the framework as <NA> internally).

Signed-off-by: Jason Dellaluce <jasondellaluce@gmail.com>