build(container): use MinGW toolchaing instead of MSVC on windows builds
The windows build broke when moving to Go 1.24, the reason is that the MSVC
toolchain linker is failing to link the static cgo library:

> libworker.a(go.o) : fatal error LNK1223: invalid or corrupt file: file contains invalid .pdata contributions [D:\a\plugins\plugins\plugins\container\build\container.vcxproj]

The issue may be related to the two different toolchains used to build
the cgo static library i.e. MinGW and the C++ code MSVC.

This change forces to use MinGW toolchain on both sides.
See also: https://github.com/golang/go/issues/20982

Signed-off-by: Iacopo Rozzo <iacopo@sysdig.com>